web/requirements.txt
Django>=1.9.7,<2.0
djangorestframework>=3.3.3,<4.0
Pillow>=3.2.0,<4.0
psycopg2>=2.6.1,<3.0
